New project (attiny1617)
The only way to change the board used in a project, is to recreate the project. This is the new project, and the old project will probably be removed sooner or later. Also very important, the pins of the dendrite have been changed in order to circumvent a strange behaviour.